requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
openai==1.54.1
gunicorn==22.0.0
